[qrcode-svg] Add documentation, change test's import
This commit adds documentation for most of the remaining things in the definition. It also changes the way the class is imported in the test, as requested by @plantain-00. This also removes the rule override in the local `tslint.json` file, to only include the extends prop. This also removes the `typeNumber` prop from options, as it doesn't actually seem to do anything in the module (it's not referenced from the options object and is just calculated during QR code creation). This also fixes a mixture of whitespace types.