Restore hls.media prop and correct type.
I erroneously removed this property in my last PR, but it's been brought to my attention by another dev. `media` is unset at construction, but is set to an `HTMLMediaElement` when passed via `attachMedia` and cleared to `null` when `detachMedia` is called. The docs appear to imply that this property is meant to be readonly, and that altering it is to be done with the above referenced methods.